refactor(ValidationViewStrategy): pull apart view strategies
The main goal of this refactoring is to pull apart the existing
ValidateCustomAttributeViewStrategy into its component parts, and
provide a means to support multiple strategies. Specifically:

  - ValidateCustomAttributeViewStrategy => ValidationViewStrategy
  - Move bootstrap specific view strategy to its own file
  - Update tests accordingly
