+++ test/validation-config.spec.js
@@ -1,5 +1,5 @@
 import {ValidationConfig} from '../src/validation/validation-config';
-import {ValidateCustomAttributeViewStrategy} from '../src/validation/validate-custom-attribute-view-strategy';
+import {TWBootstrapViewStrategy} from '../src/strategies/twbootstrap-view-strategy';
 
 describe('ValidationConfig', () => {
   it('should have default values', () => {
@@ -8,7 +8,7 @@ describe('ValidationConfig', () => {
     expect(config.getDebounceTimeout()).toBe(0);
     expect(config.getDependencies().length).toBe(0);
     expect(config.getValue('locale')).toBe('en-US');
-    expect(config.getViewStrategy()).toBe(ValidateCustomAttributeViewStrategy.TWBootstrapAppendToMessage);
+    expect(config.getViewStrategy()).toBe(TWBootstrapViewStrategy.AppendToMessage);
     expect(config.getValue('allPropertiesAreMandatory')).toBe(false);
   });
 
@@ -30,8 +30,8 @@ describe('ValidationConfig', () => {
     expect(config.getValue('locale')).toBe('nl-BE');
 
     config = new ValidationConfig();
-    expect(config.useViewStrategy(ValidateCustomAttributeViewStrategy.TWBootstrapAppendToInput)).toBe(config);// fluent API check
-    expect(config.getViewStrategy()).toBe(ValidateCustomAttributeViewStrategy.TWBootstrapAppendToInput);
+    expect(config.useViewStrategy(TWBootstrapViewStrategy.AppendToInput)).toBe(config); // fluent API check
+    expect(config.getViewStrategy()).toBe(TWBootstrapViewStrategy.AppendToInput);
 
     config = new ValidationConfig();
     expect(config.treatAllPropertiesAsMandatory()).toBe(config); // fluent API check
